+++ CHANGELOG.md
@@ -2,9 +2,11 @@
 Changes to Calva.
 
 ## [Unreleased]
+
+## [2.0.89] - 2020-03-29
 - [Add support for connecting to generic project types](https://github.com/BetterThanTomorrow/calva/issues/595)
 
-## [2.0.88] - 2020-03-21
+## [2.0.88] - 2020-03-22
 - [Change all references to `#calva-dev` so that they now point to the `#calva` Slack channel](https://clojurians.slack.com/messages/calva/)
 
 ## [2.0.87] - 2020-03-21